Moved some functions from Z2packCalculation to utils. Fixed 'merge_dict_inputs_from_parent' behavior to also do a recursive check.
