+++ aiida_z2pack/calculations/utils/__init__.py
@@ -3,3 +3,4 @@ from .prepare_pw import prepare_nscf
 from .prepare_overlap import prepare_overlap
 from .prepare_wannier90 import prepare_wannier90
 from .prepare_z2pack import prepare_z2pack
+from .utils import merge_dict_inputs_from_parent, recursive_get_linked_node, get_previous_node
